Refactor delegate normal battle classes
